Added function that fill game array with candies
It can calculate size of picture using number of columns
and line width.
It uses sweet class to create random candy objects.

I encountered many problem building this function
First was size of one cell. At first I thought it is simply
length of canvas devided by number of columns.
This calculation doesn't include uneven number of lines in the
table.
Second problem was width of the image itself and how it
correspond to length, number of columns and lineWidht.