Changed toggleSelection
Now it select only one cell.
When none has been selected before then it create first selection
When something before was selected it make sure
it isn't same cell.
If not it call function that will check if new selection is on
the north, west, east or south of previous selection.
If not it clear previous selection and create new one.