Added toggle candies function
It creates clone of gameArr and use it to find out witch
action need to be used after switching candies.

If after switch candies there are at least one set of maching
candies it calls changeCandies function.

If not it calls spinCandies function that will add animation
to previous and current candies but in the end they will come
back to it's previous position.